+++ src/Views/MovieApp/Comment.cshtml
@@ -5,20 +5,6 @@
     ViewData["Title"] =  "Home Page";
 }
 
-<script>
-  var tableBody = $('#ListComments tbody');
-  var url = '@Url.Action("buttonComment", "MovieApp/Comment")'; // adjust to suit
-  $('#buttonComment').click(function() {
-    $.get('url, function(response) {
-        tableBody.append(response);
-    });
-    }
-  });
-</script>
-
-
-<form asp-action="Comment">
-
     @if (Model !=null)
     {
     <h1>List of Comments</h1>
@@ -60,14 +46,13 @@
     <div class="form-group">
         <label asp-for="Title" class="control-label"></label>
         <input asp-for="Title" class="form-control"/>
-        <span asp-validation-for="Title" class="text-danger"></span>
+        <span asp-validation-for="Title" class="text-danger" id="comTitle"></span>
     </div>
     <div class="form-group">
         <label asp-for="Content" class="control-label"></label>
         <input asp-for="Content" class="form-control" />
-        <span asp-validation-for="Content" class="text-danger"></span>
+        <span asp-validation-for="Content" class="text-danger" id="comContent"></span>
     </div>
     <div class="form-group">
-        <input type="submit" value="Send" class="btn btn-primary" id="buttonComment" />
+        <input type="button" value="Send" class="btn btn-primary" id="buttonComment"/>
     </div>
-</form>
