+++ src/wwwroot/js/site.js
@@ -1,4 +1,14 @@
-﻿// Please see documentation at https://docs.microsoft.com/aspnet/core/client-side/bundling-and-minification
-// for details on configuring this project to bundle and minify static web assets.
-
-// Write your JavaScript code.
+﻿$(function(){
+    var tableBody = $('#ListComments tbody');
+    var url = '@Url.Action("buttonComment", "Api/V1/MovieApiController")';
+    $('#buttonComment').click(function() {
+        /*$.get(url, function(response){
+            tableBody.append(response);
+            $('#comContent').val('');
+        });*/
+        $.get(url, function(response){
+            alert('emtro');
+        });
+        alert('buton clickeado');   
+    });
+});
